Basic menu structure.
In React Native, a menu is a user interface element that provides navigation options or actions to the user. It can be a set of buttons, icons, or other interactive elements that allow the user to perform specific actions, such as navigating to different screens in the application, changing settings, performing actions related to content, among other things.